+++ Cyotek.Windows.Forms.ImageBox.Demo/DragHandleAnchor.cs
@@ -1,7 +1,7 @@
 ﻿namespace Cyotek.Windows.Forms.Demo
 {
   // Cyotek ImageBox
-  // Copyright (c) 2010-2014 Cyotek.
+  // Copyright (c) 2010-2015 Cyotek Ltd.
   // http://cyotek.com
   // http://cyotek.com/blog/tag/imagebox