+++ Cyotek.Windows.Forms.ImageBox.Demo/SwitchImageDuringZoomDemoForm.cs
@@ -7,7 +7,7 @@ using System.Windows.Forms;
 namespace Cyotek.Windows.Forms.Demo
 {
   // Cyotek ImageBox
-  // Copyright (c) 2010-2014 Cyotek.
+  // Copyright (c) 2010-2015 Cyotek Ltd.
   // http://cyotek.com
   // http://cyotek.com/blog/tag/imagebox
 
@@ -321,7 +321,7 @@ namespace Cyotek.Windows.Forms.Demo
         // actions such as modifying scrollbars occur. This means methods such as GetSourceImageRegion will return the
         // wrong X and Y values. Until this is fixed, using a timer to trigger the change.
         // However, if you had lots of map changes to make then using a timer would be a good idea regardless; for example
-        // if the user rapdily zooms through the available levels, they'll have a smoother experiance if you only load
+        // if the user rapdily zooms through the available levels, they'll have a smoother experience if you only load
         // the data once they've stopped zooming
         refreshMapTimer.Stop();
         refreshMapTimer.Start();