+++ Cyotek.Windows.Forms.ImageBox.Design/Properties/AssemblyInfo.cs
@@ -7,7 +7,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("Cyotek Ltd")]
 [assembly: AssemblyProduct("Cyotek ImageBox Control")]
-[assembly: AssemblyCopyright("Copyright © 2013-2014 Cyotek")]
+[assembly: AssemblyCopyright("Copyright © 2013-2015 Cyotek Ltd.")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 [assembly: ComVisible(false)]