+++ Cyotek.Windows.Forms.ImageBox/ImageBoxCancelEventArgs.cs
@@ -4,7 +4,7 @@ using System.Drawing;
 namespace Cyotek.Windows.Forms
 {
   // Cyotek ImageBox
-  // Copyright (c) 2010-2014 Cyotek.
+  // Copyright (c) 2010-2015 Cyotek Ltd.
   // http://cyotek.com
   // http://cyotek.com/blog/tag/imagebox