+++ Cyotek.Windows.Forms.ImageBox/ImageBoxNativeMethods.cs
@@ -8,7 +8,7 @@ using System.Runtime.InteropServices;
 namespace Cyotek.Windows.Forms
 {
   // Cyotek ImageBox
-  // Copyright (c) 2010-2014 Cyotek.
+  // Copyright (c) 2010-2015 Cyotek Ltd.
   // http://cyotek.com
   // http://cyotek.com/blog/tag/imagebox