+++ Cyotek.Windows.Forms.ImageBox/ZoomLevelCollection.cs
@@ -4,6 +4,15 @@ using System.Collections.Generic;
 
 namespace Cyotek.Windows.Forms
 {
+  // Cyotek ImageBox
+  // Copyright (c) 2010-2015 Cyotek Ltd.
+  // http://cyotek.com
+  // http://cyotek.com/blog/tag/imagebox
+
+  // Licensed under the MIT License. See license.txt for the full text.
+
+  // If you use this control in your applications, attribution, donations or contributions are welcome.
+
   /// <summary>
   /// Represents available levels of zoom in an <see cref="ImageBox"/> control
   /// </summary>